LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/charset-02-expected.txt

CONSOLE MESSAGE: SyntaxError: Invalid character '\ufffd'
CONSOLE MESSAGE: SyntaxError: Invalid character '\ufffd'

PASS UTF-8 module script with UTF-8 BOM
PASS UTF-16 module script with UTF-16BE BOM
PASS UTF-16 module script with UTF-16LE BOM
